requirements.txt
python-telegram-bot[job-queue]==21.10
anthropic>=0.42.0
httpx>=0.27.0
jinja2>=3.1.4
pydantic>=2.10.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
